requirements.txt
asgiref==3.4.1
black==22.1.0
click==8.0.4
colorama==0.4.4
Django==4.0.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
importlib-metadata==4.10.0
Markdown==3.3.6
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.5.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
zipp==3.7.0
